[tserver] KUDU-1827: tserver decommission
This change introduces two new states for TServers for decommissioning.
The states are:
	* DECOMMISSIONING_IN_PROGRESS
	* DECOMMISSIONED
In these states the TServer is quieced and new replicas are prevented to
be placed on them. When a TServer is being decommissioned, first it
enters the DECOMMISSIONING_IN_PROGRESS state, then a rebalancer job is
started to move replicas away from every TServer that is in this state.
When the rebalancing is done, the TServers move to DECOMMISSIONED state
and will stay in it until they are removed from the cluster via the 'kudu
tserver unregister' tool.
The decommissioning can be started with the following cli command:

$ kudu tserver state enter_decommissioning <master_addresses>
        <tserver_addresses> [-allow_missing_tserver]
        [-negotiation_timeout_ms=<ms>] [-timeout_ms=<ms>]

It is possible to decommission multiple TServers at once by providing
multiple addresses in a comma-separated list, which is preferred to
avoid multiple replica moves between each server decommission.

Change-Id: I15c52b653c20b2a3a45fbf8934f19f6bd1a9caea