[thirdparty] Upgrade glog to 0.6.0
Our current glog version is 0.3.5, which is very old, and apparently
broken on MacOS Monterey, as it doesn't print integers when appended to
the log stream, and it can also crash the process in certain cases.

Some of the glog API changed, requiring small changes in the code, and
the internal API, some of which Kudu relies on (stack traces,
symbolizing addresses), is now hidden, so this commit also adds a patch
to change glog's CMakeLists to make the internal API visible. Some older
patches that were merged upstream have been removed, and the one adding
stacktrace support for aarch64 had to be modified.

Change-Id: I21abd1749fdfdcde412f5a2ca2245c42da20d4f3