jwt: introduce MiniOidc
This patch takes some existing test utilities and encapsulates them into
the new MiniOidc class. The MiniOidc serves the purpose of being the
OpenID Connect Discovery endpoint, as the host for JWKSs of each
account, and the dispensor of JWTs.

This encapsulation will be useful in testing JWTs from non-C++ tests
that typically reply on exposing Mini* components via control shell.

Change-Id: I26e9b3bcd0946adbe4642a19c5ef1124e39632c6