plumb JWT authentication into clients
This change plumbs the JWT authentication into the client and into the
client negotiation (the JWTVerifier is set when building the Messenger).

There following new flags are added:
* enable_jwt_token_auth
* jwt_validate_signature (unsafe)
* jwt_allow_without_tls (unsafe)
* jwks_file_path
* jwks_url

If 'enable_jwt_token_auth' is set to true, then either 'jwks_file_path'
or 'jwks_url' has to be set, also both cannot be set at the same time.

Co-authored-by: Andrew Wong <awong@apache.org>

Change-Id: Ibdfc2640c047a2e3bb5ea03aea4439cea2532e23